Implemented Interrupt Mode
Currently, the interrupt handler keeps recalling itself, causing an infinite loop. This may be because of the built in behavior of the HAL_Master_transmit_IT command to retransmit until acknowledge is received from slave. Have to check if Noack flag is being activated causing this issue (Inside Hal drivers)